+++ weapon.h
@@ -6,28 +6,28 @@ namespace zombie {
 // Describes a Weapon and is responsible of shooting.
 class Weapon {
 public:
-	Weapon(double damage, double timeBetweenShots, double range, int clipSize) {
+	Weapon(float damage, float timeBetweenShots, float range, int clipSize) {
 		damage_ = damage;
 		timeBetweenShots_ = timeBetweenShots;
 		range_ = range;
 		clipSize_ = clipSize;
 		bulletsInWeapon_ = clipSize;
-		lastShotTime_ = 0.0;
+		lastShotTime_ = 0.0f;
 		shotSound_ = range_;
 	}
 
 	Weapon() {
-		damage_ = 35;
-		timeBetweenShots_ = 0.1;
-		range_ = 30;
+		damage_ = 35.f;
+		timeBetweenShots_ = 0.1f;
+		range_ = 30.f;
 		clipSize_ = 12;
 		bulletsInWeapon_ = 12;
-		lastShotTime_ = 0.0;
+		lastShotTime_ = 0.0f;
 		shotSound_ = range_;
 	}
 
 	// Called when user pulls the trigger.
-	bool shoot(double time) {
+	bool shoot(float time) {
 		if (lastShotTime_ + timeBetweenShots_ < time) {
 			if (bulletsInWeapon_ > 0) {
 				lastShotTime_ = time;
@@ -45,12 +45,12 @@ public:
 	}
 
 	// Returns the point damage.
-	double damage() const {
+	float damage() const {
 		return damage_;
 	}	
 
 	//
-	double range() const {
+	float range() const {
 		return range_;
 	}
 	
@@ -68,16 +68,16 @@ public:
 private:
 	bool reload_;         // Is true when the weapon is reloading.
 	
-	double damage_;  // The damage made by the weapon.
-	double timeBetweenShots_;
-	double range_;        // The range of the weapon.
+	float damage_;  // The damage made by the weapon.
+	float timeBetweenShots_;
+	float range_;        // The range of the weapon.
 	
 	int clipSize_;        // The number of bullets for a reload.
 	int bulletsInWeapon_; // The current number of bullets in the weapon.
-	double shotSound_;
+	float shotSound_;
 	
-	double timeElapsed_;  // Used as keeping track of the time.
-	double lastShotTime_;	
+	float timeElapsed_;  // Used as keeping track of the time.
+	float lastShotTime_;	
 };
 
 } // namespace zombie.
